Split src/runs/reap.ts into src/runs/reap/ subdirectory (warren-ea0c)
Step 4 of plan pl-9088. The 2114-line reap.ts + 2792-line reap.test.ts
monoliths split into per-sub-step modules + matching test files:

- reap/index.ts        — public re-exports + module doc
- reap/types.ts        — public interfaces (ReapFs, ReapExec, ReapRunInput,
                         ReapRunResult, ReapStep, ReapStepError)
- reap/run.ts          — top-level reapRun orchestrator
- reap/mulch.ts        — mulch-expertise LWW merge (+ exported mergeMulchFile)
- reap/seeds.ts        — seeds close mirror + plans mirror
- reap/plot-merge.ts   — Plot events + state replay (+ exported pure helpers)
- reap/stage.ts        — .plot/ + .seeds/ commit-through-reap
- reap/auto-plan-run.ts — frontmatter detection + dispatch helper
- reap/pr-open.ts      — auto-open PR sub-step
- reap/preview.ts      — preview launch + PR annotation sub-steps
- reap/state.ts        — failure-reason inference + terminal transition
- reap/util.ts         — splitLines, seq allocator, defaultFs/Exec
- reap/test-helpers.ts — shared test fixtures (fakeFs/Exec/BurrowClient/etc.)

Test files split alongside their modules: mulch.test.ts, plot-merge.test.ts,
run.test.ts, seeds.test.ts, state.test.ts, pr-open.test.ts, preview.test.ts,
stage.test.ts, auto-plan-run.test.ts (92 tests, all green).

Removes biome noExcessiveLinesPerFunction overrides for reap.ts and
reap.test.ts. The four functions still over the noExcessiveCognitiveComplexity
threshold (mergeMulch, mergePlot, mirrorSeeds, mirrorPlans, reapRun) carry
over their entries under their new module paths in biome.json; refactoring
those algorithms is out of scope for this mechanical split.

Removes reap.ts/reap.test.ts from scripts/file-size-budgets.json. Every new
module sits comfortably under the 500-line default threshold.

Updates consumers (src/runs/index.ts, src/runs/cancel.ts, src/runs/cancel.test.ts)
to import from ./reap/index.ts.
