+++ biome.json
@@ -64,8 +64,6 @@
 				"src/client/client.test.ts",
 				"src/preview/launch.test.ts",
 				"src/registry/refresh.test.ts",
-				"src/runs/reap.test.ts",
-				"src/runs/reap.ts",
 				"src/runs/spawn.test.ts",
 				"src/runs/stream.test.ts"
 			],
@@ -114,7 +112,10 @@
 				"src/runs/pause.ts",
 				"src/runs/pr-template.ts",
 				"src/runs/pr.ts",
-				"src/runs/reap.ts",
+				"src/runs/reap/mulch.ts",
+				"src/runs/reap/plot-merge.ts",
+				"src/runs/reap/run.ts",
+				"src/runs/reap/seeds.ts",
 				"src/runs/seed.ts",
 				"src/runs/spawn.test.ts",
 				"src/runs/spawn.ts",
